requirements.txt
streamlit 
graphql-core
requests
plotly
pytz
pandas
plyer
